requirements.txt
numpy
scikit-image
torch
torchvision
pillow
opencv-python-headless
paddlepaddle
paddlehub

